requirements.txt
django~=3.1.4
djangorestframework~=3.12.2
shapely~=1.7.1
orjson~=3.4.6
Faker~=5.0.1
psycopg2-binary~=2.8.6